+++ test/loading.test.ts
@@ -1,5 +1,24 @@
-import { expect, it } from 'vitest'
-import { Creamy } from '../src'
+import { describe, expect, it } from 'vitest'
+import { Creamy, toPascalCase } from '../src'
+
+describe('toPascalCase', () => {
+  it('should return a pascal cased string', () => {
+    expect(toPascalCase('app-button')).toBe('AppButton')
+    expect(toPascalCase('appbutton')).toBe('Appbutton')
+    expect(toPascalCase('Navbar')).toBe('Navbar')
+    expect(toPascalCase('foo bar')).toBe('FooBar')
+    expect(toPascalCase('Foo Bar')).toBe('FooBar')
+    expect(toPascalCase('fooBar')).toBe('FooBar')
+    expect(toPascalCase('FooBar')).toBe('FooBar')
+    expect(toPascalCase('foo-bar')).toBe('FooBar')
+    expect(toPascalCase('foo_bar')).toBe('FooBar')
+    expect(toPascalCase('--foo-bar--')).toBe('FooBar')
+    expect(toPascalCase('__FOO_BAR__')).toBe('FooBar')
+    expect(toPascalCase('!--foo-¿?-bar--121-*')).toBe('FooBar121')
+    expect(toPascalCase('Here i am')).toBe('HereIAm')
+    expect(toPascalCase('FOO BAR')).toBe('FooBar')
+  })
+})
 
 it('should add components', () => {
   const creamy = new Creamy()
@@ -11,7 +30,7 @@ it('should add components', () => {
 
   expect(creamy.components).toMatchInlineSnapshot(`
     Map {
-      "one" => <div>
+      "One" => <div>
         TextNode {
           "_rawText": "1",
           "childNodes": [],
@@ -21,7 +40,7 @@ it('should add components', () => {
           </div>,
         }
       </div>,
-      "two" => <div>
+      "Two" => <div>
         TextNode {
           "_rawText": "2",
           "childNodes": [],
@@ -31,7 +50,7 @@ it('should add components', () => {
           </div>,
         }
       </div>,
-      "three" => <div>
+      "Three" => <div>
         TextNode {
           "_rawText": "3",
           "childNodes": [],