+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.7.x] - 2023-09-24
+
+- Sanitize props by default 
+- Update component key and usage to be PascalCase
+
 ## [0.7.6] - 2023-09-24
 
 - Add support for `@else` and `@else-if`